Add game model, rename resetGame files
