+++ frontend/smart-kickers-game/src/App.js
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import './App.css';
 import { Button } from './components/Button';
 import GameResults from './components/GameResults.js';
-import { resetGame } from './apis/Game';
+import { resetGame } from './apis/resetGame';
 import GameStatistics from './components/GameStatistics.js';
 import { library } from '@fortawesome/fontawesome-svg-core';
 import { faPerson } from '@fortawesome/free-solid-svg-icons';
