+++ frontend/smart-kickers-game/src/apis/resetGame.js
