+++ frontend/smart-kickers-game/src/apis/resetGame.spec.js
